+++ src/router/RenovationsPaths.ts
@@ -1,4 +1,4 @@
-import { HouseAreas } from "../integration/core/constants/house-areas.constants";
+import { HouseAreas } from "../integration/core/constants/HouseAreas.constants";
 import { SubNavBarItemDTO } from "../integration/core/dtos/SubNavBarItem.dto";
 import { RenovationPathsDTO } from "../integration/core/dtos/RenovationPaths.dto";
 //TODO: Create service to fetch House areas
